Add dockerfile.
My goal in contributing to this project is to lower friction
in the development of an open server. Right now it is harder
for people to contribute, because build instructions aren't
included.

This commit moves in the direction of making it easier to contribute
by including java and a java dependency manager. In future work
I'll setup the package to be buildable using the dependency manager.

In the way of testing this commit I validated that the docker build
completed successfully.