Create index page.
This brings us a bit closer to no barrier to entry for
people to experiment with the ESF file format. Now there
is an upload form in the App that people can use to
upload ESF files and get back object files.

Validated that the index HTML was rendered via running
app.